web_services_backup.sh: Fix SimpleNote backup
The old YAML backup has long since vanished, so I have generally gone to
https://app.simplenote.com and clicked through to its "download .zip"
option. But I've now replaced the defunct URL in the backup_simplenote()
function with the (stable!) URL that redirects to, which should now
always work so long as the user is logged into SimpleNote (and they
don't change how the feature works).