cleanup-kindle.sh: Only ask once per standard ebook.
On Kindle, if a file was previously copied on and has been opened, there's a .azw3 file and a .sdr
directory. The script previously treated each individually. Now, when encountering either, if the
other is present, prompt the user to remove or keep both, and then skip the other one when it
comes up later (as it has either been already removed or explicitly kept).