general/synchronize.sh: Provide a way to run under trickle to limit bandwidth
With recent versions of Unison, I find that if a sufficiently large file
has been added to one side, attempting to transfer it over my local WiFi
network between one host with a proper WiFi card and one host that is
connected via an Ethernet-to-WiFi adapter causes the Ethernet-to-WiFi
adapter to lose its configuration (which it usually doesn't even lose
when it loses power for days or longer). (Syncing between two
natively-WiFi-connected hosts works fine.) My best guess is that the
problem is triggered by sustained high-bandwidth transmission or
receiving saturating the link, so with this I can now define
TRICKLE_MAX_KB to limit Unison's bandwidth usage.